+++ joss_paper/paper.md
@@ -52,8 +52,8 @@ Fetal brain Magnetic Resonance Imaging (MRI) is crucial for assessing neurodevel
 1. **Data Standardization**: `Fetpype` expects input data organized according to the Brain Imaging Data Structure (BIDS) standard [@gorgolewski2016brain], promoting interoperability and simplifying data management. 
 2. **Containerization**: Individual processing tools are encapsulated within Docker or Singularity containers. This ensures reproducibility and reduces installation issues, providing a better experience for the end user. 
 3. **Workflow Management**: The `Nipype` library [@gorgolewski2011nipype] is used to construct processing workflows: it provides a robust interface for combining different steps from different containers or packages, facilitating data caching and parallelization, and allowing pipelines to be easily shareable. 
-4. **Configuration**: Pipeline configuration is managed using simple YAML files and the Hydra library [@Yadan2019Hydra], allowing users to easily select between different modules or parameters without directly modifying the code. The current implementation of Fetpype integrates modules for **data preprocessing**, **super-resolution reconstruction** (NeSVoR [@xu2023nesvor], SVRTK [@kuklisova-murgasova_reconstruction_2012; @uus2022automated], or NiftyMIC [@ebner2020automated]), **segmentation** (BOUNTI [@uus2023bounti] or the developing human connectome project pipeline [@makropoulos2018developing]) and **cortical surface extraction** (using a custom implementation available at https://github.com/fetpype/surface_processing based on [@bazin2005topology; @bazin2007topology; @ma2022cortexode]).
-
+4. **Configuration**: Pipeline configuration is managed using simple YAML files and the Hydra library [@Yadan2019Hydra], allowing users to easily select between different modules or parameters without directly modifying the code. The current implementation of Fetpype integrates modules for **data preprocessing** (`Fetal-BET` [@faghihpirayesh2024fetal], non-local means denoising [@manjon2010adaptive] and N4 bias-field correction [@tustison2010n4itk]), **super-resolution reconstruction** (NeSVoR [@xu2023nesvor], SVRTK [@kuklisova-murgasova_reconstruction_2012; @uus2022automated], or NiftyMIC [@ebner2020automated]), **segmentation** (BOUNTI [@uus2023bounti] or the developing human connectome project pipeline [@makropoulos2018developing]) and **cortical surface extraction** (using a custom implementation available at https://github.com/fetpype/surface_processing based on [@bazin2005topology; @bazin2007topology; @ma2022cortexode]).
+  
 The objective underlying the conceptualization of 'Fetpype' was to provide the fetal MRI community with a standardized, reproducible, and flexible open-source platform for preprocessing and analysis. We believe this tool can facilitate research and improve the comparability across studies. We also intend to foster collaboration across research teams by providing Fetpype as a central framework that facilitates the dissemination of new image processing methods for clinical applications. The pipeline is publicly available on GitHub (https://github.com/fetpype/fetpype), and its open-source nature and modular design facilitate community involvement: researchers can integrate their own tools by creating corresponding `Nipype` interfaces and container wrappers, following the package contribution guidelines.
 
 In the future, we plan to supplement `Fetpype` with an automated reporting library containing automated quality control [@sanchez2024fetmrqc; @sanchez2025automatic], subject-wise and population-wise biometry and volumetry [@esteban2017mriqc], as well as spectral analysis [@germanaud2012larger]. We welcome contributions of authors desiring to integrate their method to `Fetpype`. 
