+++ joss_paper/paper.bib
@@ -1,3 +1,36 @@
+@article{manjon2010adaptive,
+  title     = {Adaptive non-local means denoising of MR images with spatially varying noise levels},
+  author    = {Manj{\'o}n, Jos{\'e} V and Coup{\'e}, Pierrick and Mart{\'\i}-Bonmat{\'\i}, Luis and Collins, D Louis and Robles, Montserrat},
+  journal   = {Journal of Magnetic Resonance Imaging},
+  volume    = {31},
+  number    = {1},
+  pages     = {192--203},
+  year      = {2010},
+  publisher = {Wiley Online Library},
+  doi      = {10.1002/jmri.22003}
+}
+
+@article{tustison2010n4itk,
+  title     = {{N4ITK}: improved N3 bias correction},
+  author    = {Tustison, Nicholas J and Avants, Brian B and Cook, Philip A and Zheng, Yuanjie and Egan, Alexander and Yushkevich, Paul A and Gee, James C},
+  journal   = {IEEE transactions on medical imaging},
+  volume    = {29},
+  number    = {6},
+  pages     = {1310--1320},
+  year      = {2010},
+  publisher = {IEEE},
+  doi     = {10.1109/TMI.2010.2046908}
+}
+
+@article{faghihpirayesh2024fetal,
+  title     = {{Fetal-BET}: Brain extraction tool for fetal {MRI}},
+  author    = {Faghihpirayesh, Razieh and Karimi, Davood and Erdo{\u{g}}mu{\c{s}}, Deniz and Gholipour, Ali},
+  journal   = {IEEE Open Journal of Engineering in Medicine and Biology},
+  year      = {2024},
+  publisher = {IEEE},
+  doi     = {10.1109/ojemb.2024.3426969}
+}
+
 @article{esteban2017mriqc,
   title={MRIQC: Advancing the automatic prediction of image quality in MRI from unseen sites},
   author={Esteban, Oscar and Birman, Daniel and Schaer, Marie and Koyejo, Oluwasanmi O and Poldrack, Russell A and Gorgolewski, Krzysztof J},
@@ -6,7 +39,8 @@
   number={9},
   pages={e0184661},
   year={2017},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi = {10.1371/journal.pone.0184661}
 }
 
 @article{germanaud2012larger,
@@ -17,7 +51,8 @@
   number={3},
   pages={1257--1272},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2012.07.053},
 }
 
 @article{sanchez2024fetmrqc,
@@ -27,15 +62,21 @@
   volume={97},
   pages={103282},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.media.2024.103282}
 }
 
 
-@article{sanchez2025automatic,
+@InProceedings{sanchez2025automatic,
   title={Automatic quality control in multi-centric fetal brain MRI super-resolution reconstruction},
   author={Sanchez, Thomas and Zalevskyi, Vladyslav and Mihailov, Angeline and Mart{\'\i}-Juan, Gerard and Eixarch, Elisenda and Jakab, Andras and Dunet, Vincent and Koob, M{\'e}riam and Auzias, Guillaume and Cuadra, Meritxell Bach},
-  journal={arXiv preprint arXiv:2503.10156},
-  year={2025}
+  title={Automatic Quality Control in Multi-centric Fetal Brain MRI Super-Resolution Reconstruction},
[…]
