+++ joss_paper/paper.bib
@@ -207,3 +207,12 @@
   publisher = {Elsevier},
   doi={10.1016/j.neuroimage.2019.116324}
 }
+
+@article{neves2025scanner,
+  title     = {Scanner-based real-time automated volumetry reporting of the fetus, amniotic fluid, placenta, and umbilical cord for fetal MRI at 0.55 T},
+  author    = {Neves Silva, Sara and Uus, Alena and Waheed, Hadi and Bansal, Simi and St Clair, Kamilah and Norman, Wendy and Aviles Verdera, Jordina and Cromb, Daniel and Woodgate, Tomas and Van Poppel, Milou and others},
+  journal   = {Magnetic resonance in medicine},
+  year      = {2025},
+  publisher = {Wiley Online Library},
+  doi = {10.1002/mrm.70097}
+}
