+++ joss_paper/paper.md
@@ -19,6 +19,14 @@ authors:
   - name: David Meunier
     orcid: 0000-0002-5812-6138
     affiliation: 4
+  - name: Miguel Angel Gonzalez Ballester
+    orcid: 0000-0002-9227-6826
+    equal-contrib: true
+    affiliation: "3, 5" # (Multiple affiliations must be quoted)
+  - name: Oscar Camara
+    orcid: 0000-0002-5125-6132
+    equal-contrib: true
+    affiliation: 3 # (Multiple affiliations must be quoted)
   - name: Gemma Piella
     orcid: 0000-0001-5236-5819
     affiliation: 3
@@ -37,13 +45,16 @@ affiliations:
    index: 3
  - name: Aix-Marseille Université, CNRS, Institut de Neurosciences de La Timone, France
    index: 4
-date: 19 June 2025
+ - name: ICREA, Barcelona, Spain
+   index: 5
+ 
+date: 7 October 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-Fetal brain Magnetic Resonance Imaging (MRI) is crucial for assessing neurodevelopment *in utero*. However, analyzing this data presents significant challenges due to fetal motion, low signal-to-noise ratio, and the need for complex multi-step processing, including motion correction, super-resolution reconstruction, and segmentation. While various specialized tools exist for individual steps, integrating them into robust, reproducible, and user-friendly workflows that go from raw image to volume and surface analysis is not straightforward. This lack of standardization hinders reproducibility across studies and limits the adoption of advanced analysis techniques for researchers and clinicians. To address these challenges, we introduce `Fetpype`, an open-source Python library designed to streamline and standardize the preprocessing and analysis of T2-weighted fetal brain MRI data as illustrated in (\autoref{fig:fetpype}).
+Fetal brain Magnetic Resonance Imaging (MRI) is crucial for assessing neurodevelopment *in utero*. However, analyzing this data presents significant challenges due to fetal motion, low signal-to-noise ratio, and the need for complex multi-step processing, including motion correction, super-resolution reconstruction, segmentation, and surface extraction. While various specialized tools exist for individual steps, integrating them into robust, reproducible, and user-friendly workflows that go from raw images to processed volumes is not straightforward. This lack of standardization hinders reproducibility across studies and limits the adoption of advanced analysis techniques for researchers and clinicians. To address these challenges, we introduce `Fetpype`, an open-source Python library designed to streamline and standardize the preprocessing and analysis of T2-weighted fetal brain MRI data as illustrated in (\autoref{fig:fetpype}).
 
 # Statement of need
 
@@ -52,16 +63,21 @@ Fetal brain Magnetic Resonance Imaging (MRI) is crucial for assessing neurodevel
 1. **Data Standardization**: `Fetpype` expects input data organized according to the Brain Imaging Data Structure (BIDS) standard [@gorgolewski2016brain], promoting interoperability and simplifying data management. 
 2. **Containerization**: Individual processing tools are encapsulated within Docker or Singularity containers. This ensures reproducibility and reduces installation issues, providing a better experience for the end user. 
 3. **Workflow Management**: The `Nipype` library [@gorgolewski2011nipype] is used to construct processing workflows: it provides a robust interface for combining different steps from different containers or packages, facilitating data caching and parallelization, and allowing pipelines to be easily shareable. 
[…]
