+++ joss_paper/paper.bib
@@ -76,7 +76,7 @@
   publisher={Springer Nature Switzerland},
   address={Cham},
   pages={3--14},
-  doi={https://link.springer.com/chapter/10.1007/978-3-032-05997-0_1}
+  doi={10.1007/978-3-032-05997-0_1}
 }
 
 @article{ma2022cortexode,
