### 2025-12-12 8fc18c6
add fetalsynthseg support for segmetnation option
