### 2025-12-12 5b53cb7
revert back to bounti as default seg configs/default_docker.yaml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
