### 2026-01-22 e1c9509
Updated paper to accomodate changes in JOSS requirements.
