+++ joss_paper/paper.md
@@ -21,12 +21,13 @@ authors:
     affiliation: 4
   - name: Miguel Angel Gonzalez Ballester
     orcid: 0000-0002-9227-6826
-    equal-contrib: true
     affiliation: "3, 5" # (Multiple affiliations must be quoted)
   - name: Oscar Camara
     orcid: 0000-0002-5125-6132
-    equal-contrib: true
     affiliation: 3 # (Multiple affiliations must be quoted)
+  - name: Elisenda Eixarch 
+    orcid: 0000-0001-7379-9608
+    affiliation: "6, 7" # (Multiple affiliations must be quoted)
   - name: Gemma Piella
     orcid: 0000-0001-5236-5819
     affiliation: 3
@@ -47,18 +48,30 @@ affiliations:
    index: 4
  - name: ICREA, Barcelona, Spain
    index: 5
- 
-date: 7 October 2025
+ - name: BCNatal | Fetal Medicine Research Center (Hospital Clínic and Hospital Sant Joan de Déu, Universitat de Barcelona), Barcelona, Spain
+   index: 6
+ - name: Institut d’Investigacions Biomèdiques August Pi i Sunyer (IDIBAPS), Barcelona, Spain and Centre for Biomedical Research on Rare Diseases (CIBERER), Barcelona, Spain
+   index: 7
+date: 22 January 2026
 bibliography: paper.bib
 ---
 
+
 # Summary
 
-Fetal brain Magnetic Resonance Imaging (MRI) is crucial for assessing neurodevelopment *in utero*. However, analyzing this data presents significant challenges due to fetal motion, low signal-to-noise ratio, and the need for complex multi-step processing, including motion correction, super-resolution reconstruction, segmentation, and surface extraction. While various specialized tools exist for individual steps, integrating them into robust, reproducible, and user-friendly workflows that go from raw images to processed volumes is not straightforward. This lack of standardization hinders reproducibility across studies and limits the adoption of advanced analysis techniques for researchers and clinicians. To address these challenges, we introduce `Fetpype`, an open-source Python library designed to streamline and standardize the preprocessing and analysis of T2-weighted fetal brain MRI data as illustrated in (\autoref{fig:fetpype}).
+Fetal brain magnetic resonance imaging (MRI) is crucial for assessing neurodevelopment *in utero*. However, fetal MRI analysis remains technically challenging due to fetal motion, low signal-to-noise ratio, and the need for complex multi-step processing pipelines. These pipelines typically include motion correction, super-resolution reconstruction, tissue segmentation, and cortical surface extraction. While specialized tools exist for each individual processing step, integrating them into a robust, reproducible, and user-friendly end-to-end workflow remains difficult. This fragmentation limits reproducibility across studies and hinders the adoption of advanced fetal neuroimaging methods in both research and clinical contexts.
+
+`Fetpype` addresses this gap by providing a standardized, modular, and reproducible framework for fetal brain MRI preprocessing and analysis, enabling researchers to process raw T2-weighted acquisitions through to derived volumetric and surface-based outputs within a unified workflow.
 
 # Statement of need
+`Fetpype` is an open-source Python package designed to streamline and standardize the preprocessing and analysis of T2-weighted fetal brain MRI data. The package targets the fetal neuroimaging community, where methodological heterogeneity and complex software dependencies have historically limited reproducibility and comparability across studies.
 
-`Fetpype` is a Python package integrating several established neuroimaging software principles and tools to create a cohesive and extensible framework, which we summarize in four points. 
[…]
