### 2026-01-28 dd3beac
Update JOSS paper + added reference to the arxiv paper.
