+++ joss_paper/paper.md
@@ -89,7 +89,7 @@ The overall processing workflow is summarized in \autoref{fig:fetpype}.
 # Research impact statement
 `Fetpype` is the result of a longstanding collaboration within a European consortium of researchers specializing in fetal neuroimaging. Its default configurations and processing workflows have been the result of extensive testing to achieve robust processing on data acquired across multiple hospitals in France, Spain, and Switzerland, covering a range of scanners and acquisition protocols.
 
-The framework has been used to process large-scale fetal MRI datasets within the consortium, has contributed to a first publication (citation pending), and is supporting ongoing research projects. Fetpype is used by multiple research groups and has begun to receive external contributions, including pull requests that integrate additional processing methods. This suggests that `Fetpype` addresses a clear methodological need and can serve as shared community infrastructure for fetal brain MRI research.
+The framework has been used to process large-scale fetal MRI datasets within the consortium, has contributed to a first publication [@sanchez2026data], and is supporting ongoing research projects. Fetpype is used by multiple research groups and has begun to receive external contributions, including pull requests that integrate additional processing methods. This suggests that `Fetpype` addresses a clear methodological need and can serve as shared community infrastructure for fetal brain MRI research.
 
 
 In the future, we plan to supplement `Fetpype` with an automated reporting library containing automated quality control [@sanchez2024fetmrqc; @sanchez2025automatic], subject-wise and population-wise biometry and volumetry [@esteban2017mriqc; @neves2025scanner], as well as spectral analysis of surfaces [@germanaud2012larger]. We welcome community contributions, particularly implementations of new methods that can be integrated into the existing containerized workflow framework.
