+++ README.md
@@ -12,7 +12,17 @@ It aims at making easy the integration of a variety of tools existing to process
 
 More information are available on the [documentation](https://fetpype.github.io/fetpype/).
 
+# Reference
+If you use `Fetpype` in your research, please cite the following paper:
 
+```
+@article{sanchez2025fetpype,
+  title={Fetpype: An open-source pipeline for reproducible fetal brain mri analysis},
+  author={Sanchez, Thomas and Mart{\'\i}-Juan, Gerard and Meunier, David and Ballester, Miguel Angel Gonzalez and Camara, Oscar and Piella, Gemma and Cuadra, Meritxell Bach and Auzias, Guillaume},
+  journal={arXiv preprint arXiv:2512.17472},
+  year={2025}
+}
+```
 
 
 
