### 2026-01-28 #9214 @t-sanchez
Thanks a lot for heads up @rkurchin. We believe that our article fits in the scope of JOSS even with the changes, and we updated the paper accordingly.
