### 2026-02-13 #9214 @jromanowska
@drmowinckels, 👋 , how is it going? I was wondering whether you have time to review this submission to JOSS? You can read about the review process [in JOSS docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
If you cannot take it on this time, perhaps you can recommend someone who might fit?
