### 2026-02-13 #9214 @t-sanchez
Hi @jromanowska,
Thanks a lot for taking care of the process! Looking at the similar papers published at JOSS, it looks like people like robbisg, PeerHerholz, TheChymera, emdupre could be good candidate reviewers.
