### 2026-03-05 #9214 @t-sanchez
Hi @jromanowska,

I've been looking through similar papers to ours on JOSS, and I wanted to suggest other potentiel reviewers, if this can help: djmannion, htwangtw, 62442katieb, grlee77 could be options, in addition to the ones I had suggested previously: robbisg and PeerHerholz.

If there's anything that I can do/people that I can contact review, let me know 😄
