### 2026-03-16 #9214 @jromanowska
:wave: @62442katieb, @djmannion, @PeerHerholz & @robbisg, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
