### 2026-03-24 #10278 @jromanowska
👋🏼 @t-sanchez, @dtamadauw, @mathieuboudreau, @rserial, this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

I encourage you to **over-communicate** and let everyone know that you're on the task every now and then, instead of waiting several weeks to collect all your comments, questions, or suggestions.

_Please feel free to ping me (@jromanowska) if you have any questions/concerns._
