### 2026-04-07 #10278 @jromanowska
@mathieuboudreau, @rserial, just a reminder to generate your checklist. This will help you get started and will enable the authors and me to follow as you go through the points there and check them off.
Also: if you know that you need some more time, please, write here.
