### 2026-04-07 #10278 @dtamadauw
I have completed a review of the repository and the paper. Actually, I enjoyed it very much as the repository is well-organized. The software is structured very well and addresses a significant need in the fetal MRI community.

I'm very impressed by the community of fetal brain MRI. In many MRI sub-fields, organizers often struggle to find a sufficient number of contributors because the communities are basically very small. However, this repository successfully attracts many contributors and is actually used among them. I can definitely say this software deserves more than acceptance into JOSS. The software concept is understandable and reasonable. Because fragmented packages often provide challenges in quantitative MRI in general, their approach enables stable and low-barrier analysis for multi-site and multi-vendor studies.

I have 3 comments to improve the repository before acceptance:

1. Add citation.cff and References: I did not see a dedicated references section in the paper. Please confirm the paper includes the appropriate citations and consider adding a citation.cff file to the repository.

2. Detailed Requirements: Please update the requirements section to explicitly list system-level dependencies, specifically Docker/Singularity and Graphviz.

3. Robust Testing: I had an OSError: No command 'dot' found during the pipeline tests. It looks that Graphviz is a system-level requirement for running the full test, but it's not currently listed in the installation instructions or requirements.txt. I recommend adding a note to the documentation for Linux/macOS users to install the Graphviz binary. Also, please consider wrapping write_graph in a try-except block or making it an optional test to prevent failures in environments without Graphviz. By the way, I tested on Ubuntu 24.04, worked perfectly with Graphviz.
