### 2026-04-10 fetpype/fetpype#99 @mathieuboudreau
Installation: Specify range of compatible Python versions (>= or =< >=)
In your installation documentation,

> Within your desired python environment, install fetpype

```pip install -e .```

isn't descriptive enough, because your [dependencies](https://github.com/fetpype/fetpype/blob/f641925e01e27ee2dce81069d8660909092f0c28/pyproject.toml#L13-L18) won't be compatible with all python versions (you typically can view on pipy their range). For ease and better communication, you can simply define the earliest Python version you've tested the software on and confirmed that it worked, and state that in your documentation. Or better yet, you can also enforce this in your pyproject file, see here: https://ericmjl.github.io/blog/2023/7/12/how-to-configure-a-minimum-python-version-in-pyprojecttoml/

Mainly, this is important because otherwise, you'll get some issues opened by users indicating they can't install it, and will end up repeating this conversation with multiple users. It's also better for long-term reproducibility, but not perfect (because, you only depend the minimum package dependency versions, in the future they'll have major API changes and/or update their own python compatibility limits), but defining a minimum is a good first step.
