### 2026-04-10 fetpype/fetpype#99 @mathieuboudreau
I was able to successfully run `pip install -e .` for this project on the latest release available on conda (`python 3.14.4`), however for the earliest available release (`python 3.8.5`) it doesn't install:

```
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Getting requirements to build editable ... done
  Preparing editable metadata (pyproject.toml) ... done
INFO: pip is looking at multiple versions of fetpype to determine which version is compatible with other requirements. This could take a while.
ERROR: Ignored the following versions that require a different python version: 1.10.0 Requires-Python >=3.9; 1.11.0 Requires-Python >=3.10; 1.9.0 Requires-Python >=3.9; 1.9.1 Requires-Python >=3.9; 1.9.2 Requires-Python >=3.9
ERROR: Could not find a version that satisfies the requirement nipype==1.10.0 (from fetpype) (from versions: 0.3.4, 0.4, 0.4.1, 0.5.1, 0.5.2, 0.5.3, 0.6.0, 0.7.0, 0.8.0, 0.9.0, 0.9.1, 0.9.2, 0.10.0, 0.11.0rc1, 0.11.0, 0.12.0rc1, 0.12.0rc2, 0.12.0, 0.12.1, 0.13.0rc1, 0.13.0, 0.13.1, 0.14.0rc1, 0.14.0, 1.0.0, 1.0.1, 1.0.2, 1.0.3, 1.0.4, 1.1.0, 1.1.1, 1.1.2, 1.1.3, 1.1.4, 1.1.5, 1.1.6, 1.1.7, 1.1.8, 1.1.9, 1.2.0, 1.2.1, 1.2.2, 1.2.3, 1.3.0rc1, 1.3.0, 1.3.1, 1.3.2, 1.4.0, 1.4.1, 1.4.2, 1.5.0rc1, 1.5.0, 1.5.1, 1.6.0, 1.6.1, 1.7.0, 1.7.1, 1.8.0, 1.8.1, 1.8.2, 1.8.3, 1.8.4, 1.8.5, 1.8.6)
ERROR: No matching distribution found for nipype==1.10.0
(py385) mathieuboudreau@Mathieus-MacBook-Pro fetpype %
```

So there needs to be a hard minimum limit defined preferable in both the pyproject file and documentation
