### 2026-04-10 fetpype/fetpype#100 @mathieuboudreau
Installation: Add graphviz as a dependency?
To ensure my installation ran correctly, I tried running your test suite, however many failed:

```

t_surf>'}, 'surface_lh': {'use_scheme': 'bounti', 'out_file': 'hemi-L_white.surf.gii', 'labelling_scheme': {'bounti': [5, 7, 14, 16]}}, 'surface_rh': {'use_scheme': 'bounti', 'out_file': 'hemi-R_white.surf.gii', 'labelling_scheme': {'bounti': [6, 8, 15, 17]}}}, 'container': 'docker', 'save_graph': True}
--------------------------- Captured stdout teardown ---------------------------
Mock output dir /private/var/folders/l1/1xswjffd73l8yp7dd7pq9lyw0000gn/T/pytest-of-mathieuboudreau/pytest-0/output23 will be cleaned up by pytest.
================================ tests coverage ================================
_______________ coverage: platform darwin, python 3.14.4-final-0 _______________

Name                       Stmts   Miss  Cover
----------------------------------------------
tests/__init__.py              0      0   100%
tests/conftest.py             62     30    52%
tests/test_pipelines.py       27      1    96%
tests/test_utils_bids.py      72      1    99%
----------------------------------------------
TOTAL                        161     32    80%
=========================== short test summary info ============================
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_full_pipeline_sr-niftymic_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_full_pipeline_sr-niftymic_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_full_pipeline_sr-nesvor_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_full_pipeline_sr-nesvor_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_full_pipeline_sr-svrtk_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_full_pipeline_sr-svrtk_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_rec_pipeline_sr-niftymic_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_rec_pipeline_sr-niftymic_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_rec_pipeline_sr-nesvor_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_rec_pipeline_sr-nesvor_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_rec_pipeline_sr-svrtk_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_rec_pipeline_sr-svrtk_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_seg_pipeline_sr-niftymic_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_seg_pipeline_sr-niftymic_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_seg_pipeline_sr-nesvor_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_seg_pipeline_sr-nesvor_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_seg_pipeline_sr-svrtk_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_seg_pipeline_sr-svrtk_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_surf_pipeline_sr-niftymic_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_surf_pipeline_sr-niftymic_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_surf_pipeline_sr-nesvor_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_surf_pipeline_sr-nesvor_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_surf_pipeline_sr-svrtk_loadMask-True] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
FAILED tests/test_pipelines.py::test_create_full_pipelines[pipelines-create_surf_pipeline_sr-svrtk_loadMask-False] - OSError: No command "dot" found on host Mathieus-MacBook-Pro.local. Please ...
======================== 24 failed, 10 passed in 12.70s ========================
(py3144) mathieuboudreau@Mathieus-MacBook-Pro fetpype %

```

I installed graphviz, and now the tests pass for me:

```

(py3144) mathieuboudreau@Mathieus-MacBook-Pro fetpype % py.test --cov tests --ignore=initial_scripts/
=============================================================================================== test session starts ================================================================================================
platform darwin -- Python 3.14.4, pytest-9.0.3, pluggy-1.6.0
rootdir: /Users/mathieuboudreau/neuropoly/reviews/fetpype
configfile: pyproject.toml
plugins: cov-7.1.0, hydra-core-1.3.2
collected 34 items

tests/test_pipelines.py ...........................                                                                                                                                                          [ 79%]
tests/test_utils_bids.py .......                                                                                                                                                                             [100%]

================================================================================================== tests coverage ==================================================================================================
_________________________________________________________________________________ coverage: platform darwin, python 3.14.4-final-0 _________________________________________________________________________________

Name                       Stmts   Miss  Cover
----------------------------------------------
tests/__init__.py              0      0   100%
tests/conftest.py             62     30    52%
tests/test_pipelines.py       27      0   100%
tests/test_utils_bids.py      72      1    99%
----------------------------------------------
TOTAL                        161     31    81%
=============================================================================================== 34 passed in 10.12s ================================================================================================
(py3144) mathieuboudreau@Mathieus-MacBook-Pro fetpype %

```
