### 2026-04-10 #10278 @mathieuboudreau
For **Functionality --> Installation**, I've opened issues that should be resolved prior to ticking off the box:

- [ ] https://github.com/fetpype/fetpype/issues/99
- [ ] https://github.com/fetpype/fetpype/issues/100
- [ ] https://github.com/fetpype/fetpype/issues/102
