### 2026-04-10 fetpype/fetpype#102 @mathieuboudreau
Installation: Missing Docker
Your code assumes that Docker is already installed, but the [installation](https://fetpype.github.io/fetpype/#installation) documentation gives no direction to the new user that they need to install it or where to find documentation on how to do so.

A new user that followed your installation instructions will encounter this when they first try to use your suggested command:

```

mathieuboudreau@Mathieus-MacBook-Pro fetpype % fetpype_run --data ./bids --out ./output --config ./configs/default_docker.yaml
Initialising the pipeline...
preprocessing:
  brain_extraction:
    docker:
      cmd: docker run --gpus all <mount> fetpype/fetpype_utils:latest run_brain_extraction
        --input_stacks <input_stacks> --output_masks <output_masks> --method fet_bet
    singularity:
      cmd: singularity run --bind <singularity_mount> --nv <singularity_path>/fetpype_utils.sif
        run_brain_extraction --input_stacks <input_stacks> --output_masks <output_masks>
        --method monaifbs
  check_stacks_and_masks:
    enabled: true
  denoising:
    enabled: true
    docker:
      cmd: docker run <mount> fetpype/fetpype_utils:0.1.2 run_denoising --input_stacks
        <input_stacks> --output_stacks <output_stacks>
    singularity:
      cmd: singularity run --bind <singularity_mount> <singularity_path>/fetpype_utils.sif
        run_denoising --input_stacks <input_stacks> --output_stacks <output_stacks>
  cropping:
    enabled: true
  bias_correction:
    enabled: true
    docker:
      cmd: docker run <mount> fetpype/fetpype_utils:latest run_bias_field_correction
        --input_stacks <input_stacks> --input_masks <input_masks> --output_stacks
        <output_stacks>
    singularity:
      cmd: singularity run --bind <singularity_mount> <singularity_path>/fetpype_utils.sif
        run_bias_field_correction --input_stacks <input_stacks> --input_masks <input_masks>
        --output_stacks <output_stacks>
reconstruction:
  pipeline: nesvor
  docker:
    cmd: docker run --gpus '"device=0"' <mount> junshenxu/nesvor:v0.5.0 nesvor reconstruct
      --input-stacks <input_stacks> --stack-masks <input_masks> --output-volume <output_volume>
      --batch-size 4096 --n-levels-bias 1
  singularity:
    cmd: singularity exec --bind <singularity_mount> --nv <singularity_path>/nesvor.sif
      nesvor reconstruct --input-stacks <input_stacks> --stack-masks <input_masks>
      --output-volume <output_volume> --batch-size 4096 --n-levels-bias 1
  args:
    path_to_output: nesvor.nii.gz
  output_resolution: 0.8
segmentation:
  pipeline: bounti
  docker:
    cmd: docker run --rm <mount> fetalsvrtk/segmentation:general_auto_amd bash /home/auto-proc-svrtk/scripts/auto-brain-bounti-segmentation-fetal.sh
      <input_dir> <output_dir>
  singularity:
    cmd: singularity exec --bind <singularity_mount> --home <singularity_home> --nv
      <singularity_path>/bounti.sif bash /home/auto-proc-svrtk/scripts/auto-brain-bounti-segmentation-fetal.sh
      <input_dir> <output_dir>
  path_to_output: <basename>-mask-brain_bounti-19.nii.gz
surface:
  pipeline: surfpype
  docker:
    cmd: docker run --rm <mount> fetpype/surf_proc:0.0.2 generate_mesh -l <labelling_scheme>
      -s <input_seg> -m <output_surf>
  singularity:
    cmd: singularity exec --bind <singularity_mount> --home <singularity_home> --nv
      <singularity_path>/surf_proc.sif generate_mesh -l <labelling_scheme> -s <input_seg>
      -m <output_surf>
  surface_lh:
    use_scheme: bounti
    out_file: hemi-L_white.surf.gii
    labelling_scheme:
      bounti:
      - 5
      - 7
      - 14
      - 16
  surface_rh:
    use_scheme: bounti
    out_file: hemi-R_white.surf.gii
    labelling_scheme:
      bounti:
      - 6
      - 8
      - 15
      - 17
container: docker
save_graph: true

▶ nesvor_bounti_surfpype.input.bids_datasource [sub-01]
260410-12:52:16,712 nipype.workflow INFO:
	 [Node] Setting-up "nesvor_bounti_surfpype.input.bids_datasource" in "/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/input/_subject_01/bids_datasource".
260410-12:52:16,714 nipype.workflow INFO:
	 [Node] Executing "bids_datasource" <nipype.interfaces.io.BIDSDataGrabber>
260410-12:52:16,830 nipype.workflow INFO:
	 [Node] Finished "bids_datasource", elapsed time 0.100006s.
✔ nesvor_bounti_surfpype.input.bids_datasource [sub-01] (2.0s)
▶ nesvor_bounti_surfpype.full_pipeline.Preprocessing.BrainExtraction [sub-01]
260410-12:52:17,592 nipype.workflow INFO:
	 [Node] Setting-up "nesvor_bounti_surfpype.full_pipeline.Preprocessing.BrainExtraction" in "/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction".
260410-12:52:17,596 nipype.workflow INFO:
	 [Node] Executing "BrainExtraction" <nipype.interfaces.utility.wrappers.Function>
260410-12:52:17,601 nipype.workflow INFO:
	 Running: docker run --gpus all -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat:/Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output:/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output fetpype/fetpype_utils:latest run_brain_extraction --input_stacks /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat/sub-01_run-1_T2w.nii.gz --output_masks /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output/sub-01_run-1_mask.nii.gz --method fet_bet
docker: Cannot connect to the Docker daemon at unix:///Users/mathieuboudreau/.docker/run/docker.sock. Is the docker daemon running?.
See 'docker run --help'.
260410-12:52:17,654 nipype.workflow INFO:
	 [Node] Finished "BrainExtraction", elapsed time 0.055589s.
260410-12:52:17,654 nipype.workflow WARNING:
	 Storing result file without outputs
260410-12:52:17,656 nipype.workflow WARNING:
	 [Node] Error on "nesvor_bounti_surfpype.full_pipeline.Preprocessing.BrainExtraction" (/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction)
✖ nesvor_bounti_surfpype.full_pipeline.Preprocessing.BrainExtraction [sub-01] failed (see crashfile & logs)
Node BrainExtraction.a0 failed to run on host Mathieus-MacBook-Pro.local.
Saving crash info to /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/logs/20260410-125213/crash-20260410-125219-mathieuboudreau-BrainExtraction.a0-f4c89910-bfcc-4ff5-8ca3-d0a4f383b99f.txt
Traceback (most recent call last):
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/plugins/multiproc.py", line 67, in run_node
    result["result"] = node.run(updatehash=updatehash)
                       ~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 525, in run
    result = self._run_interface(execute=True)
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 643, in _run_interface
    return self._run_command(execute)
           ~~~~~~~~~~~~~~~~~^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 769, in _run_command
    raise NodeExecutionError(msg)
nipype.pipeline.engine.nodes.NodeExecutionError: Exception raised while executing Node BrainExtraction.

Traceback:
	Traceback (most recent call last):
	  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/interfaces/base/core.py", line 401, in run
	    runtime = self._run_interface(runtime)
	  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/interfaces/utility/wrappers.py", line 139, in _run_interface
	    out = function_handle(**args)
	  File "<string>", line 107, in run_prepro_cmd
	  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/utils/logging.py", line 283, in run_and_tee
	    raise RuntimeError(
	    ...<3 lines>...
	    )
	RuntimeError: Docker call failed with exit code 125.
	Command: docker run --gpus all -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat:/Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output:/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output fetpype/fetpype_utils:latest run_brain_extraction --input_stacks /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat/sub-01_run-1_T2w.nii.gz --output_masks /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output/sub-01_run-1_mask.nii.gz --method fet_bet
	Output:
	docker: Cannot connect to the Docker daemon at unix:///Users/mathieuboudreau/.docker/run/docker.sock. Is the docker daemon running?.See 'docker run --help'.

could not run node: nesvor_bounti_surfpype.full_pipeline.Preprocessing.BrainExtraction.a0
Uncaught exception
Traceback (most recent call last):
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/bin/fetpype_run", line 6, in <module>
    sys.exit(main())
             ~~~~^^
  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/workflows/pipeline_fet.py", line 277, in main
    create_main_workflow(
    ~~~~~~~~~~~~~~~~~~~~^
        data_dir=args.data,
        ^^^^^^^^^^^^^^^^^^^
    ...<10 lines>...
        verbose=args.verbose,
        ^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/workflows/pipeline_fet.py", line 253, in create_main_workflow
    main_workflow.run(
    ~~~~~~~~~~~~~~~~~^
        plugin="MultiProc",
        ^^^^^^^^^^^^^^^^^^^
        plugin_args={"n_procs": nprocs, "status_callback": status_line},
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/workflows.py", line 625, in run
    runner.run(execgraph, updatehash=updatehash, config=self.config)
    ~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/plugins/base.py", line 223, in run
    raise error from cause
RuntimeError: Traceback (most recent call last):
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/plugins/multiproc.py", line 67, in run_node
    result["result"] = node.run(updatehash=updatehash)
                       ~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 525, in run
    result = self._run_interface(execute=True)
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 643, in _run_interface
    return self._run_command(execute)
           ~~~~~~~~~~~~~~~~~^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 769, in _run_command
    raise NodeExecutionError(msg)
nipype.pipeline.engine.nodes.NodeExecutionError: Exception raised while executing Node BrainExtraction.

Traceback:
	Traceback (most recent call last):
	  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/interfaces/base/core.py", line 401, in run
	    runtime = self._run_interface(runtime)
	  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/interfaces/utility/wrappers.py", line 139, in _run_interface
	    out = function_handle(**args)
	  File "<string>", line 107, in run_prepro_cmd
	  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/utils/logging.py", line 283, in run_and_tee
	    raise RuntimeError(
	    ...<3 lines>...
	    )
	RuntimeError: Docker call failed with exit code 125.
	Command: docker run --gpus all -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat:/Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output:/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output fetpype/fetpype_utils:latest run_brain_extraction --input_stacks /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat/sub-01_run-1_T2w.nii.gz --output_masks /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output/sub-01_run-1_mask.nii.gz --method fet_bet
	Output:
	docker: Cannot connect to the Docker daemon at unix:///Users/mathieuboudreau/.docker/run/docker.sock. Is the docker daemon running?.See 'docker run --help'.

Traceback (most recent call last):
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/bin/fetpype_run", line 6, in <module>
    sys.exit(main())
             ~~~~^^
  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/workflows/pipeline_fet.py", line 277, in main
    create_main_workflow(
    ~~~~~~~~~~~~~~~~~~~~^
        data_dir=args.data,
        ^^^^^^^^^^^^^^^^^^^
    ...<10 lines>...
        verbose=args.verbose,
        ^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/workflows/pipeline_fet.py", line 253, in create_main_workflow
    main_workflow.run(
    ~~~~~~~~~~~~~~~~~^
        plugin="MultiProc",
        ^^^^^^^^^^^^^^^^^^^
        plugin_args={"n_procs": nprocs, "status_callback": status_line},
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/workflows.py", line 625, in run
    runner.run(execgraph, updatehash=updatehash, config=self.config)
    ~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/plugins/base.py", line 223, in run
    raise error from cause
RuntimeError: Traceback (most recent call last):
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/plugins/multiproc.py", line 67, in run_node
    result["result"] = node.run(updatehash=updatehash)
                       ~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 525, in run
    result = self._run_interface(execute=True)
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 643, in _run_interface
    return self._run_command(execute)
           ~~~~~~~~~~~~~~~~~^^^^^^^^^
  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/pipeline/engine/nodes.py", line 769, in _run_command
    raise NodeExecutionError(msg)
nipype.pipeline.engine.nodes.NodeExecutionError: Exception raised while executing Node BrainExtraction.
Traceback:
	Traceback (most recent call last):
	  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/interfaces/base/core.py", line 401, in run
	    runtime = self._run_interface(runtime)
	  File "/Users/mathieuboudreau/mambaforge/envs/py3144/lib/python3.14/site-packages/nipype/interfaces/utility/wrappers.py", line 139, in _run_interface
	    out = function_handle(**args)
	  File "<string>", line 107, in run_prepro_cmd
	  File "/Users/mathieuboudreau/neuropoly/reviews/fetpype/fetpype/utils/logging.py", line 283, in run_and_tee
	    raise RuntimeError(
	    ...<3 lines>...
	    )
	RuntimeError: Docker call failed with exit code 125.
	Command: docker run --gpus all -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat:/Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat -v /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output:/Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output fetpype/fetpype_utils:latest run_brain_extraction --input_stacks /Users/mathieuboudreau/neuropoly/reviews/fetpype/bids/sub-01/anat/sub-01_run-1_T2w.nii.gz --output_masks /Users/mathieuboudreau/neuropoly/reviews/fetpype/output/nipype/nesvor_bounti_surfpype/full_pipeline/Preprocessing/_subject_01/BrainExtraction/output/sub-01_run-1_mask.nii.gz --method fet_bet
	Output:
	docker: Cannot connect to the Docker daemon at unix:///Users/mathieuboudreau/.docker/run/docker.sock. Is the docker daemon running?.See 'docker run --help'.
(py3144) mathieuboudreau@Mathieus-MacBook-Pro fetpype %

```
