### 2026-04-10 fetpype/fetpype#103 @mathieuboudreau
Functionality: Missing instruction to start Docker prior to running commants
Even with Docker installed (which I did), a new user will encounter the error described in #102 if they haven't started it on their computer prior to running your pipeline command. Direction should be given to the user somewhere; as this is something that needs to be done each time and not just during the installtion, it should probably be instructed near the CLI commands sections [here](https://fetpype.github.io/fetpype/#running-your-first-pipeline) or [here](https://fetpype.github.io/fetpype/run_parts/)
