### 2026-04-10 #10278 @mathieuboudreau
I'm finished my pass, when the four items I opened issues for are resolved (i.e. issues are closed), I'll check off the final review boxes.

Great work overall - thank you for contributing this to the community!
