### 2026-04-10 #10278 @gauzias
Hi everyone,
many thanks @mathieuboudreau and @dtamadauw for the very constructive reviews, and @jromanowska for managing the process.
We will work as soon as possible on the issues that @mathieuboudreau did open, and on the 3 aspects pointed by @dtamadauw.
These improvements will be beneficial both for future users as well as for us to maintain the repos, so very helpful!
This is our first experience with JOSS and we appreciate a lot the process!
