### 2026-04-13 6575e00
Add a dynamically generated path to the config instead of hardcoding it.
