### 2026-04-13 a4a6b85
Address issues #100, #102, #103 by documenting the Graphviz requirements and expliciting that Docker or Singularity must be installed and running prior to running fetpype.
