### 2026-04-13 #10278 @rserial
@jromanowska yes I do, I have a couple of deadlines but I should be able to go through it on Wednesday. Would that still be ok?
