### 2026-04-13 b110dc3
Explicit python>=3.9 in the docs and update readme as it mentioned that surface extraction was to be implemented, but has been implemented for a while.
