### 2026-04-13 10d3cf3
Fix typo in index.md
