### 2026-04-13 fetpype/fetpype#99 @t-sanchez
Hi @mathieuboudreau

Thanks for the issue. We've addressed this in #104 where we have:
1. Added the requirement of `python>=3.9` in `pyproject.toml` https://github.com/fetpype/fetpype/blob/9e0626256952f5a6372c711e9c0857b07efca84f/pyproject.toml#L8
2. Explicited this requirement in the docs
https://github.com/fetpype/fetpype/blob/10d3cf3806e516fa45c6dd2274fccc730fe59ea6/docs/index.md?plain=1#L44

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
