### 2026-04-13 ccf5a72
Fix formatting issue in index.md
