### 2026-04-13 fetpype/fetpype#102 @t-sanchez
Hi @mathieuboudreau

Thanks for the issue. We've addressed this in #104 where we have explicitly mentioned that Docker should be installed and running before any `Fetpype` command is issued. We've pointed users to the Docker installation website as well.

https://github.com/fetpype/fetpype/blob/f28e6c73fd940ed1553e515bbe88257bfa72a5e8/docs/index.md?plain=1#L42-L47

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
