### 2026-04-13 fetpype/fetpype#103 @t-sanchez
Hi @mathieuboudreau

Thanks for the issue. As with #102, we've addressed this in #104. We have mentioned to users that they can run `docker info` to make sure that Docker is running prior to issuing `Fetpype` command.

https://github.com/fetpype/fetpype/blob/f28e6c73fd940ed1553e515bbe88257bfa72a5e8/docs/index.md?plain=1#L42-L47

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
