### 2026-04-13 #10278 @t-sanchez
Hi @dtamadauw, @mathieuboudreau

Thank you for the comments! We have addressed them with https://github.com/fetpype/fetpype/pull/104.

For @mathieuboudreau, I answered directly on your issues, if this is OK 🙂

For @dtamadauw, I'll just briefly comment here!

> Add citation.cff and References: I did not see a dedicated references section in the paper. Please confirm the paper includes the appropriate citations and consider adding a citation.cff file to the repository.

Thanks for the suggestion of adding a `citation.cff`. It's not something I was familiar with but this is indeed very nice :) We have added it! Regarding the references, in `paper.md`, it's actually normal for the markdown file to not have references. They are in the [paper.bib](https://github.com/fetpype/fetpype/blob/main/joss_paper/paper.bib) file and are added by the engine when building the paper.

> Detailed Requirements: Please update the requirements section to explicitly list system-level dependencies, specifically Docker/Singularity and Graphviz.

Good idea, thank you! @mathieuboudreau also suggested this so we have added this in the documentation here:
https://github.com/fetpype/fetpype/blob/f28e6c73fd940ed1553e515bbe88257bfa72a5e8/docs/index.md?plain=1#L42-L47

> Robust Testing: I had an OSError: No command 'dot' found during the pipeline tests. It looks that Graphviz is a system-level requirement for running the full test, but it's not currently listed in the installation instructions or requirements.txt. I recommend adding a note to the documentation for Linux/macOS users to install the Graphviz binary. Also, please consider wrapping write_graph in a try-except block or making it an optional test to prevent failures in environments without Graphviz. By the way, I tested on Ubuntu 24.04, worked perfectly with Graphviz.

As mentioned above, we have now mentioned `Graphviz` as a requirement. Regarding the tests, as our main use is running them through automated CI like Github actions, that we install graphviz in this environement (cf. the [workflow yaml file](https://github.com/fetpype/fetpype/blob/f28e6c73fd940ed1553e515bbe88257bfa72a5e8/.github/workflows/check_on_PR.yml#L22-L23)) and that we mention now `Graphviz` as a requirement, we believe that this fits well our intended use! But if this is still an issue, we could of course implement this try/except blocks in the tests.

Thanks again for the positive reviews and I hope that our changes address well your comments!
