### 2026-04-16 fetpype/fetpype#105 @rserial
pytest missing from dev dependencies
when trying to run the automated tests for the project, 'pytest' wasnt included in the pyproject.toml or any requirements file.
It should be added to the pyproject dependencies so it can be installed easily with pip intall -e (as described in documentation).
