### 2026-04-16 fetpype/fetpype#106 @rserial
Add graphviz dependency missing
graphviz is a system-level requirement but is not listed in pyproject.toml. The current documentation only covers debian-based systems (sudo apt-get), leaving out other platforms like macos. Given the tool's focus on containerization, it may also be worh considering whether this dependency could be handled inside the docker container or replaced by the pypi graphviz package.
