### 2026-04-16 fetpype/fetpype#107 @rserial
missing example usage data (functionality test)
The repo lacks a toy dataset (e.g. hosted on Zenodo) to test the functionality of the library. There is no example tutorial, quick script or anything to help users to learn how to use the library or verify their results are correct.
