### 2026-04-16 fetpype/fetpype#108 @rserial
missing documentation on how to run automated tests
Although the project does have automated tests implemented, neither the contributing and readme files explain how to run the tests locally.
A short section in the contributing documentation would fix this.
