### 2026-04-16 #10278 @rserial
Hi all, I've completed my first review round. I created a couple of issues related to installation and functionality.

**Installation + documentation**
- [ ]  [pytest missing from dev dependencies](https://github.com/fetpype/fetpype/issues/105)

- [ ] [Add graphviz dependency missing](https://github.com/fetpype/fetpype/issues/106)

- [ ]  [missing documentation on how to run automated tests](https://github.com/fetpype/fetpype/issues/108)

**Functionality + documentation**
I was actually looking forward trying the library myself but I couldn't find any data to do so. This may be because I'm not from the fetal MRI field, but a toy dataset and a short tutorial would help here. It would also help me verify whether the library performs as expected, since the results in the paper looked quite promissing.

- [ ] [missing example usage data (functionality test)](https://github.com/fetpype/fetpype/issues/107)
