### 2026-04-20 #10278 @jromanowska
Thank you, @rserial. As you can see from the discussion above, some of the issues you found are already taken care of.

@mathieuboudreau, @dtamadauw, do you have further comments to the authors?
