### 2026-04-27 136ca09
Describe graphviz installation beyond ubuntu; add pytest as a dev requirement + update pyproject.toml.
