### 2026-04-27 5fd7ac9
Fixed issues regarding a run using pre-set masks (needed for the example run with FaBiAN data)
