+++ fetpype/pipelines/full_pipeline.py
@@ -470,7 +470,8 @@ def create_full_pipeline(cfg, load_masks=False, name="full_pipeline"):
 
     # PREPROCESSING
     full_fet_pipe.connect(inputnode, "stacks", prepro_pipe, "inputnode.stacks")
-
+    if load_masks:
+        full_fet_pipe.connect(inputnode, "masks", prepro_pipe, "inputnode.masks")
     # RECONSTRUCTION
     full_fet_pipe.connect(
         prepro_pipe, "outputnode.stacks", recon, "inputnode.stacks"
@@ -545,6 +546,7 @@ def create_rec_pipeline(cfg, load_masks=False, name="rec_pipeline"):
     rec_pipe.connect(inputnode, "stacks", prepro_pipe, "inputnode.stacks")
     if load_masks:
         rec_pipe.connect(inputnode, "masks", prepro_pipe, "inputnode.masks")
+
     # RECONSTRUCTION
 
     rec_pipe.connect(
