### 2026-04-27 b5cb967
Add test data as well as a tutorial to the docs.
