### 2026-04-27 e24d14d
Remove debugging logging
