+++ fetpype/pipelines/full_pipeline.py
@@ -471,7 +471,7 @@ def create_full_pipeline(cfg, load_masks=False, name="full_pipeline"):
     # PREPROCESSING
     full_fet_pipe.connect(inputnode, "stacks", prepro_pipe, "inputnode.stacks")
     if load_masks:
-        full_fet_pipe.connect(inputnode, "masks", 
+        full_fet_pipe.connect(inputnode, "masks",
                               prepro_pipe, "inputnode.masks")
     # RECONSTRUCTION
     full_fet_pipe.connect(
