### 2026-04-27 fetpype/fetpype#106 @t-sanchez
Hi @rserial,

Thanks for the issue. We addressed this in #109, where we explicited the installation for macOS and pointed to the graphviz website for other operating systems.

https://github.com/fetpype/fetpype/blob/4f3461091a377bb9495c6858095a07e3fa5a4e58/docs/index.md?plain=1#L46-L49

Sadly, given the that `nipype` uses graphviz and is needed to run the core of the fetpype library, it would not be straightforward to have a containerization of graphviz. Also, the pypi graphviz package is, if I'm not mistaken, a wrapper that calls onto a system dependency, which would not address the issue.

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
