### 2026-04-27 fetpype/fetpype#105 @t-sanchez
Hi @rserial,

Thanks for the issue. We have addressed this in #109, where we included pytest as a dev dependency in `pyproject.toml`.

https://github.com/fetpype/fetpype/blob/4f3461091a377bb9495c6858095a07e3fa5a4e58/pyproject.toml#L21-L25

We also added a section in `contributing.md` highlighting how to install the `dev` environment and run the tests.

https://github.com/fetpype/fetpype/blob/4f3461091a377bb9495c6858095a07e3fa5a4e58/docs/contributing.md?plain=1#L21-L30

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
