### 2026-04-27 fetpype/fetpype#107 @t-sanchez
Hi @rserial,

Thanks for the issue, this is a very nice idea! We've addressed this in #109.

What we did is actually add a small toy simulated dataset directly into the repo (~1MB) under `test_data` that can be used in the [tutorial](https://fetpype.github.io/fetpype/tutorial/) that we added in the documentation. This includes visual results to illustrate how the outputs of fetpype look.

Note that given that fetal brain data are inherently sensitive data, it is difficult to share actual raw data. This is why we shared simulated data, with the downside that the simulation does *not* model the maternal tissue around the brain. As a result, we must use pre-computed masks as the brain extraction step would otherwise fail. This was something already covered by fetpype, and we discuss this in the tutorial.

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
