### 2026-04-27 fetpype/fetpype#108 @t-sanchez
Hi @rserial,

As mentioned in #105, we have added this in the `contributing.md` file [here](https://fetpype.github.io/fetpype/contributing/).

Let us know if we should do further changes and feel free to close the issue otherwise 🙂
