### 2026-04-27 #10278 @t-sanchez
Hi @rserial,

Thanks for the comments! We have addressed them in [#109](https://github.com/fetpype/fetpype/pull/109), and we have answered directly on the issues 🙂

The comment regarding the data is especially helpful. It's true that fetal data is scarce, as releasing these data openly is difficult due to their sensitive nature. We have added a tutorial and a small simulated dataset (single subject) on which fetpype can be run as a demo!

I hope that this can address your comments.
