### 2026-05-17 fetpype/fetpype#107 @rserial
Hi @t-sanchez, thanks a lot for the changes made, especially including simulated data to test the output of the library and the tutorial.
I cloned the new version of the repo including the test data and everything in the tutorial seems to run well (just a small issue [#110](https://github.com/fetpype/fetpype/issues/110)), till unfortunately the denoising step crashes. I get the following errors:

``` Error in denoising -- Skipping the stack sub-simu001_ses-01_run-1_T2w.nii.gz: Command '['DenoiseImage', '-i', ...]' died with <Signals.SIGILL: 4>  ```

Despite this, the process continues till the next step (bias correction), which then crashes because the denoising output files were never created (I think):

``` Error in bias correction -- Skipping the stack sub-simu001_ses-01_run-1_T2w.nii.gz: [...] sitk::ERROR: The file ".../Denoising/mapflow/_Denoising0/output/sub-simu001_ses-01_run-1_T2w.nii.gz" does not exist. ```

I was wondering if these errors could be related to the fact that the Docker image is compiled for linux/amd64 and I'm running it on a Mac with Apple Silicon (ARM)? I checked my Docker desktop configuration and I have the Rosetta emulation enabled. If that is the case, and the Apple Silicon Macs are not supported, that is totally fine but it should be mentioned in the documentation.
