### 2026-05-18 c905060
Edit docs to clarify that fetpype currently does not run on apple silicon
