### 2026-05-18 fetpype/fetpype#107 @t-sanchez
Hi @rserial,

Thanks a lot for finding this out. Indeed, Fetpype currently does not run on Apple Silicon macs because the dockers are build for amd, and ANTs crashes even with Rosetta emulation.

I have updated the documentation to reflect this here : https://github.com/fetpype/fetpype/blob/26ceef9c6ad174fde78eadf597050b4f7f6d3052/docs/index.md?plain=1#L53-L54
