+++ citation.cff
@@ -1,8 +1,8 @@
 cff-version: 1.2.0                                                                                                                                                                                                                                    
 message: "If you use this software, please cite it as below."                                                                                                                                                                                         
 title: "Fetpype: An Open-Source Pipeline for Reproducible Fetal Brain MRI Analysis"                                                                                                                                                                   
-version: "1.0.0"                                                                                                                                                                                                                                      
-date-released: 2026-01-22                                                                                                                                                                                                                             
+version: "v1.0.1"                                                                                                                                                                                                                                      
+date-released: 2026-05-19                                                                                                                                                                                                                             
 url: "https://github.com/fetpype/fetpype"                                                                                                                                                                                                             
 authors:                                                                                                                                                                                                                                              
   - family-names: Sanchez                                                                                                                                                                                                                             
