### 2026-05-20 c867c42
Final changes in the JOSS paper
