+++ joss_paper/paper.md
@@ -66,7 +66,7 @@ Fetal brain magnetic resonance imaging (MRI) is crucial for assessing neurodevel
 # Statement of need
 `Fetpype` is an open-source Python package designed to streamline and standardize the preprocessing and analysis of T2-weighted fetal brain MRI data. The package targets the fetal neuroimaging community, where methodological heterogeneity and complex software dependencies have historically limited reproducibility and comparability across studies.
 
-Existing fetal brain MRI tools typically focus on individual processing steps and require customized code for pre- and post-processing, as well as to connect different modules, making it difficult to reproduce processing results across studies. `Fetpype` addresses these challenges by providing a configurable, containerized, and `Nipype`-driven solution that integrates state-of-the-art fetal MRI processing tools into a cohesive pipeline. By emphasizing reproducibility, extensibility, and ease of use, `Fetpype` lowers the barrier to applying advanced fetal MRI analysis methods and facilitates consistent processing across sites, scanners, and studies.  In doing so, `Fetpype` improves comparability across studies and supports community collaboration by facilitating the dissemination of new image processing methods for clinical applications. The pipeline is publicly available on \href{https://github.com/fetpype/fetpype}{GitHub}.
+Existing fetal brain MRI tools typically focus on individual processing steps and require customized code for pre- and post-processing, as well as to connect different modules, making it difficult to reproduce processing results across studies. `Fetpype` addresses these challenges by providing a configurable, containerized, and `Nipype`-driven solution that integrates state-of-the-art fetal MRI processing tools into a cohesive pipeline. By emphasizing reproducibility, extensibility, and ease of use, `Fetpype` lowers the barrier to applying advanced fetal MRI analysis methods and facilitates consistent processing across sites, scanners, and studies.  In doing so, `Fetpype` improves comparability across studies and supports community collaboration by facilitating the dissemination of new image processing methods for clinical applications. The pipeline is publicly available on \href{https://github.com/fetpype/fetpype}{GitHub} [@fetpype_v1.0.1].
 
 ![The different steps covered by `Fetpype`. Starting from several T2-weighted stacks of thick slices of the fetal brain (_acquisition_), `Fetpype` pre-processes data before feeding them to a _super-resolution reconstruction_ algorithm that fuses them in a single high-resolution volume. This volume then undergoes _segmentation_, before moving to cortical _surface extraction_. \label{fig:fetpype}](fetpype.png)
 
