+++ biblio.bib
@@ -1,0 +1,1685 @@
+@article{lajous2022fetal,
+  title     = {A fetal brain magnetic resonance acquisition numerical phantom (FaBiAN)},
+  author    = {Lajous, Helene and Roy, Christopher W and Hilbert, Tom and de Dumast, Priscille and Tourbier, Sebastien and Aleman-Gomez, Yasser and Yerly, Jerome and Yu, Thomas and Kebiri, Hamza and Payette, Kelly and others},
+  journal   = {Scientific Reports},
+  volume    = {12},
+  number    = {1},
+  pages     = {8682},
+  year      = {2022},
+  publisher = {Nature Publishing Group UK London}
+}
+
+@article{manjon2010adaptive,
+  title={Adaptive non-local means denoising of MR images with spatially varying noise levels},
+  author={Manj{\'o}n, Jos{\'e} V and Coup{\'e}, Pierrick and Mart{\'\i}-Bonmat{\'\i}, Luis and Collins, D Louis and Robles, Montserrat},
+  journal={Journal of Magnetic Resonance Imaging},
+  volume={31},
+  number={1},
+  pages={192--203},
+  year={2010},
+  publisher={Wiley Online Library}
+}
+
+@article{tustison2010n4itk,
+  title={{N4ITK}: improved N3 bias correction},
+  author={Tustison, Nicholas J and Avants, Brian B and Cook, Philip A and Zheng, Yuanjie and Egan, Alexander and Yushkevich, Paul A and Gee, James C},
+  journal={IEEE transactions on medical imaging},
+  volume={29},
+  number={6},
+  pages={1310--1320},
+  year={2010},
+  publisher={IEEE}
+}
+
+@article{faghihpirayesh2024fetal,
+  title={{Fetal-BET}: Brain extraction tool for fetal {MRI}},
+  author={Faghihpirayesh, Razieh and Karimi, Davood and Erdo{\u{g}}mu{\c{s}}, Deniz and Gholipour, Ali},
+  journal={IEEE Open Journal of Engineering in Medicine and Biology},
+  year={2024},
+  publisher={IEEE}
+}
+
+@article{payette2024multi,
+  title     = {Multi-center fetal brain tissue annotation (FeTA) challenge 2022 results},
+  author    = {Payette, Kelly and others},
+  journal   = {IEEE Transactions on Medical Imaging},
+  year      = {2024},
+  publisher = {IEEE}
+}
+%Steger, C{\'e}line and Licandro, Roxane and De Dumast, Priscille and Li, Hongwei Bran and Barkovich, Matthew and Li, Liu and Dannecker, Maik and Chen, Chen and Ouyang, Cheng and others
+@inproceedings{de2021segmentation,
+  title        = {Segmentation of the cortical plate in fetal brain MRI with a topological loss},
+  author       = {de Dumast, Priscille and Kebiri, Hamza and Atat, Chirine and Dunet, Vincent and Koob, M{\'e}riam and Cuadra, Meritxell Bach},
+  booktitle    = {Uncertainty for Safe Utilization of Machine Learning in Medical Imaging, and Perinatal Imaging, Placental and Preterm Image Analysis: 3rd International Workshop, UNSURE 2021, and 6th International Workshop, PIPPI 2021, Held in Conjunction with MICCAI 2021, Strasbourg, France, October 1, 2021, Proceedings 3},
+  pages        = {200--209},
+  year         = {2021},
+  organization = {Springer}
+}
+
+@article{zha2025data,
+  title     = {Data-centric artificial intelligence: A survey},
+  author    = {Zha, Daochen and others},
+  journal   = {ACM Computing Surveys},
+  volume    = {57},
+  number    = {5},
+  pages     = {1--42},
+  year      = {2025},
+  publisher = {ACM New York, NY}
+}
[…]
