### 2026-07-16 c421054
Update PR: modular postprocessing with quantile_ratio in the file itself, put back generic path in template SG config
