+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "fetpype"
-version = "1.0.0"
+version = "1.0.2"
 requires-python = ">=3.9"
 description = "Utilities for preprocessing fetal brain MRI scans"
 authors = [{name = "Thomas Sanchez", email = "thomas.sanchez@unil.ch"},
