+++ configs/reconstruction/niftymic.yaml
@@ -17,20 +17,3 @@ reconstruction:
       --dir-output <output_dir>"
 path_to_output: "recon_template_space/srr_template.nii.gz"
   
-postprocessing:
-  clamp_intensity:
-    enabled: false
-  bias_correction:
-    enabled: false
-    docker:
-      cmd: "docker run <mount> fetpype/fetpype_utils:latest run_bias_field_correction 
-        --input_stacks <input_stacks> 
-        --input_masks <input_masks> 
-        --output_stacks <output_stacks>"
-    singularity:
-      cmd: "singularity run --bind <singularity_mount>
-        <singularity_path>/fetpype_utils.sif
-        run_bias_field_correction
-        --input_stacks <input_stacks> 
-        --input_masks <input_masks> 
-        --output_stacks <output_stacks>"
