+++ configs/reconstruction/svrtk.yaml
@@ -12,20 +12,4 @@ reconstruction:
       <input_dir> <output_dir> 1 <input_tp> <output_res> 1"
 path_to_output: "reo-SVR-output-brain.nii.gz"
 
-postprocessing:
-  clamp_intensity:
-    enabled: false
-  bias_correction:
-    enabled: false
-    docker:
-      cmd: "docker run <mount> fetpype/fetpype_utils:latest run_bias_field_correction 
-        --input_stacks <input_stacks> 
-        --input_masks <input_masks> 
-        --output_stacks <output_stacks>"
-    singularity:
-      cmd: "singularity run --bind <singularity_mount>
-        <singularity_path>/fetpype_utils.sif
-        run_bias_field_correction
-        --input_stacks <input_stacks> 
-        --input_masks <input_masks> 
-        --output_stacks <output_stacks>"
+
