+++ configs/default_sg.yaml
@@ -1,15 +1,14 @@
 defaults:
   - preprocessing/default
   - reconstruction/nesvor
+  - reconstruction/postprocessing
   - segmentation/bounti
   - surface/surfpype
   - _self_
 container: "singularity"
-singularity_path: /scratch/dborooghani/softs/fetpype_singularity_images # e.g. "/home/gmarti/SINGULARITY"
-singularity_mount: /scratch/dborooghani:/scratch/dborooghani,/scratch/gauzias:/scratch/gauzias
-singularity_home: /scratch/dborooghani/softs/tmp_fetpype:/home/tmp_proc/ # e.g. "/gpfs/home/gmarti/tmp_proc"
+singularity_path: <singularity_base> # e.g. "/home/gmarti/SINGULARITY"
+singularity_mount: <mounting> # e.g. "/gpfs/home/gmarti:/gpfs/home/gmarti"
+singularity_home: <home folder> # e.g. "/gpfs/home/gmarti/tmp_proc"
 reconstruction:
   output_resolution: 0.5
-  quantile_ratio: 0.997
 save_graph: False
-
