+++ configs/segmentation/fetalsynthseg.yaml
@@ -6,6 +6,6 @@ docker:
     python3 predict.py
     --model drifts
     --input <input_dir>/input_srr.nii.gz 
-    --output <output_dir>/seg-drifts_pred.nii.gz
+    --output <output_dir>/seg-fetalsynthseg_pred.nii.gz
     --gpu --lateralize"
-path_to_output: "seg-drifts_pred.nii.gz"
+path_to_output: "seg-fetalsynthseg_pred.nii.gz"
