+++ configs/surface/surfpype.yaml
@@ -18,9 +18,11 @@ surface_lh:
     out_file: "hemi-L_white.surf.gii"
     labelling_scheme:
         bounti: [5, 7, 14, 16]
+        fetalsynthseg: [13, 14, 16]
 
 surface_rh:
     use_scheme: "bounti"
     out_file: "hemi-R_white.surf.gii"
     labelling_scheme:
         bounti: [6, 8, 15, 17]
+        fetalsynthseg: [3, 4, 6]
