### 2026-07-16 43ff41c
update tests to include postprocessing
