+++ tests/configs/default_sg.yaml
@@ -1,6 +1,7 @@
 defaults:
   - preprocessing/default
   - reconstruction/niftymic
+  - reconstruction/postprocessing
   - segmentation/bounti
   - surface/surface
   - _self_
