### 2026-07-16 16544b8
update citation cff for JOSS
