### 2026-07-16 d199494
Update tests to also test fetalsynthseg
