+++ fetpype/utils/utils_bids.py
@@ -118,7 +118,8 @@ def create_datasource(
     for sub in subjects:
         if sub not in existing_sub:
             raise ValueError(
-                f"Requested subject {sub} was not found in the " f"folder {data_dir}."
+                f"Requested subject {sub} was not found in the "
+                f"folder {data_dir}."
             )
 
         existing_ses = layout.get_sessions(subject=sub)
@@ -130,7 +131,9 @@ def create_datasource(
         sessions_subj = [None] if len(sessions_subj) == 0 else sessions_subj
         for ses in sessions_subj:
             if ses is not None and ses not in existing_ses:
-                print(f"WARNING: Session {ses} was not found for subject {sub}.")
+                print(
+                    f"WARNING: Session {ses} was not found for subject {sub}."
+                )
             existing_acq = layout.get_acquisition(subject=sub, session=ses)
             if acquisitions is None:
                 acquisitions_subj = existing_acq
@@ -176,7 +179,9 @@ def create_bids_datasink(
     <datatype>/<BIDS_filename>
     """
     if not strip_dir:
-        raise ValueError("`strip_dir` (Nipype work dir base path) is required.")
+        raise ValueError(
+            "`strip_dir` (Nipype work dir base path) is required."
+        )
     if name is None:
         name = f"{pipeline_name}_datasink"
 
@@ -322,7 +327,11 @@ def create_bids_datasink(
             )
         )
     # ** Rule 4.5: Segmentation Output FetalSynthSeg **
-    if seg_label == "fetalsynthseg" and rec_label and pipeline_name != "preprocessing":
+    if (
+        seg_label == "fetalsynthseg"
+        and rec_label
+        and pipeline_name != "preprocessing"
+    ):
         # with session
         regex_subs.append(
             (
@@ -447,7 +456,9 @@ def create_bids_datasink(
     return datasink
 
 
-def create_datasink(iterables, name="output", params_subs={}, params_regex_subs={}):
+def create_datasink(
+    iterables, name="output", params_subs={}, params_regex_subs={}
+):
     """
     Deprecated. Creates a data sink node for reformatting and organizing
     relevant outputs.
@@ -498,7 +509,9 @@ def create_datasink(iterables, name="output", params_subs={}, params_regex_subs=
     datasink.inputs.substitutions = subjFolders
 
     # Load regex-based substitutions from the 'regex_subs.json' file
-    json_regex_subs = op.join(op.dirname(op.abspath(__file__)), "regex_subs.json")
+    json_regex_subs = op.join(
+        op.dirname(op.abspath(__file__)), "regex_subs.json"
+    )
     dict_regex_subs = json.load(open(json_regex_subs, encoding="utf-8"))
 
     # Update with provided regex substitutions
@@ -540,9 +553,13 @@ def get_gestational_age(bids_dir, T2):
             df["participant_id"] == f"{subject_id}", "gestational_age"
         ].values[0]
     except KeyError:
-        raise KeyError("Column 'gestational_age' not found in participants.tsv")
+        raise KeyError(
+            "Column 'gestational_age' not found in participants.tsv"
+        )
     except IndexError:
-        raise IndexError(f"Subject sub-{subject_id} not found in participants.tsv")
+        raise IndexError(
+            f"Subject sub-{subject_id} not found in participants.tsv"
+        )
 
     return gestational_age
 
