### 2026-07-16 c0451df
Update contribution guide to have mkdocs and flake8
