### 2026-07-17 26c0612
Revert the default docker and SG script to use bounti instead of fetalsynthseg
