### 2026-07-17 1a33ddb
add postprocessing in marseille example
