+++ configs/example_sg_marseille.yaml
@@ -1,6 +1,7 @@
 defaults:
   - preprocessing/default
   - reconstruction/nesvor
+  - reconstruction/postprocessing
   - segmentation/bounti
   - surface/surfpype
   - _self_
