### 2026-07-17 61ac385
add singularity command
