+++ configs/segmentation/fetalsynthseg.yaml
@@ -8,4 +8,11 @@ docker:
     --input <input_dir>/input_srr.nii.gz 
     --output <output_dir>/seg-fetalsynthseg_pred.nii.gz
     --gpu --lateralize"
+singularity:
+  cmd: "singularity exec <mount> --nv fetalsynthseg.sif \
+    python3 predict.py \
+    --model drifts \
+    --input <input_dir>/input_srr.nii.gz \
+    --output <output_dir>/seg-fetalsynthseg_pred.nii.gz \
+    --gpu --lateralize"
 path_to_output: "seg-fetalsynthseg_pred.nii.gz"
