### 2026-07-20 e85aae1
update singularity command
