### 2026-07-23 4958578
add --pwd flag to sg cmd
