+++ configs/segmentation/fetalsynthseg.yaml
@@ -9,9 +9,9 @@ docker:
     --output <output_dir>/seg-fetalsynthseg_pred.nii.gz
     --gpu --lateralize"
 singularity:
-  cmd: "singularity exec --bind <singularity_mount> --nv
+  cmd: "singularity exec --pwd /fetalSynthSeg --bind <singularity_mount> --nv
     <singularity_path>/fetalsynthseg.sif
-    python3 /fetalSynthSeg/predict.py
+    python3 predict.py
     --model drifts
     --input <input_dir>/input_srr.nii.gz
     --output <output_dir>/seg-fetalsynthseg_pred.nii.gz
