+++ biblio.bib
@@ -1,0 +1,1704 @@
+@article{zalevskyi2026evaluating,
+  title   = {Evaluating Synthetic Data Generation for Domain Generalization in Fetal Brain MRI Segmentation},
+  author  = {Zalevskyi, Vladyslav and Sanchez, Thomas and Roulet, Margaux and Bulut, Busra and Lajous, Hélène and Aviles Verdera, Jordina and Neves Silva, Sara and Langs, Georg and Kasprian, Gregor and Licandro, Roxane and others},
+  journal = {Machine Learning for Biomedical Imaging},
+  volume  = {2026},
+  number  = {July 2026 issue},
+  pages   = {482--506},
+  year    = {2026}
+}
+
+@inproceedings{zalevskyi2024improving,
+  title        = {Improving cross-domain brain tissue segmentation in fetal MRI with synthetic data},
+  author       = {Zalevskyi, Vladyslav and Sanchez, Thomas and Roulet, Margaux and Aviles Verdera, Jordina and Hutter, Jana and Kebiri, Hamza and Bach Cuadra, Meritxell},
+  booktitle    = {International Conference on Medical Image Computing and Computer-Assisted Intervention},
+  pages        = {437--447},
+  year         = {2024},
+  organization = {Springer}
+}
+
+@article{lajous2022fetal,
+  title     = {A fetal brain magnetic resonance acquisition numerical phantom (FaBiAN)},
+  author    = {Lajous, Helene and Roy, Christopher W and Hilbert, Tom and de Dumast, Priscille and Tourbier, Sebastien and Aleman-Gomez, Yasser and Yerly, Jerome and Yu, Thomas and Kebiri, Hamza and Payette, Kelly and others},
+  journal   = {Scientific Reports},
+  volume    = {12},
+  number    = {1},
+  pages     = {8682},
+  year      = {2022},
+  publisher = {Nature Publishing Group UK London}
+}
+
+@article{manjon2010adaptive,
+  title={Adaptive non-local means denoising of MR images with spatially varying noise levels},
+  author={Manj{\'o}n, Jos{\'e} V and Coup{\'e}, Pierrick and Mart{\'\i}-Bonmat{\'\i}, Luis and Collins, D Louis and Robles, Montserrat},
+  journal={Journal of Magnetic Resonance Imaging},
+  volume={31},
+  number={1},
+  pages={192--203},
+  year={2010},
+  publisher={Wiley Online Library}
+}
+
+@article{tustison2010n4itk,
+  title={{N4ITK}: improved N3 bias correction},
+  author={Tustison, Nicholas J and Avants, Brian B and Cook, Philip A and Zheng, Yuanjie and Egan, Alexander and Yushkevich, Paul A and Gee, James C},
+  journal={IEEE transactions on medical imaging},
+  volume={29},
+  number={6},
+  pages={1310--1320},
+  year={2010},
+  publisher={IEEE}
+}
+
+@article{faghihpirayesh2024fetal,
+  title={{Fetal-BET}: Brain extraction tool for fetal {MRI}},
+  author={Faghihpirayesh, Razieh and Karimi, Davood and Erdo{\u{g}}mu{\c{s}}, Deniz and Gholipour, Ali},
+  journal={IEEE Open Journal of Engineering in Medicine and Biology},
+  year={2024},
+  publisher={IEEE}
+}
+
+@article{payette2024multi,
+  title     = {Multi-center fetal brain tissue annotation (FeTA) challenge 2022 results},
+  author    = {Payette, Kelly and others},
+  journal   = {IEEE Transactions on Medical Imaging},
+  year      = {2024},
+  publisher = {IEEE}
+}
[…]
