### 2026-07-31 67cfcb0
add note for mesocentre users
