+++ fetpype/utils/utils_bids.py
@@ -207,7 +207,7 @@ def create_bids_datasink(
                         rf"^{escaped_bids_derivatives_root}/"
                         rf".*?_?session_([^/]+)"
                         rf"_subject_([^/]+).*?/?_denoising.*/"
-                        rf"(sub-[^_]+_ses-[^_]+(?:_run-\d+))?"
+                        rf"(sub-[^_]+_ses-[^_]+(?:_acq-[^_]+)?(?:_run-\d+))?"
                         rf"_T2w_noise_corrected(\.nii(?:\.gz)?)$"
                     ),
                     (
@@ -223,7 +223,7 @@ def create_bids_datasink(
                         rf"^{escaped_bids_derivatives_root}/"
                         rf"(?!.*?_?session_[^/]+).*?_?subject_([^/]+).*?/"
                         rf"?_denoising.*/"
-                        rf"(sub-[^_]+(?:_run-\d+)?)?_?"
+                        rf"(sub-[^_]+(?:_acq-[^_]+)?(?:_run-\d+)?)?_?"
                         rf"T2w_noise_corrected(\.nii(?:\.gz)?)$"
                     ),
                     (
@@ -241,7 +241,7 @@ def create_bids_datasink(
                     (
                         rf"^{escaped_bids_derivatives_root}/.*?_session_"
                         rf"([^/]+)_subject_([^/]+).*/?_cropping.*/"
-                        rf"(sub-[^_]+_ses-[^_]+(?:_run-\d+)?)_"
+                        rf"(sub-[^_]+_ses-[^_]+(?:_acq-[^_]+)?(?:_run-\d+)?)_"
                         rf"mask(\.nii(?:\.gz)?)$"
                     ),
                     (
@@ -257,7 +257,7 @@ def create_bids_datasink(
                         rf"^{escaped_bids_derivatives_root}/"
                         rf"(?!.*?_session_[^/]+).*?_subject_"
                         rf"([^/]+).*/?_cropping.*/"
-                        rf"(sub-[^_]+(?:_run-\d+)?)_mask(\.nii(?:\.gz)?)$"
+                        rf"(sub-[^_]+(?:_acq-[^_]+)?(?:_run-\d+)?)_mask(\.nii(?:\.gz)?)$"
                     ),
                     (
                         rf"{bids_derivatives_root}/sub-\1/{datatype}"
@@ -268,12 +268,27 @@ def create_bids_datasink(
 
     # ** Rule 3: Reconstruction Output **
     if rec_label and not seg_label and pipeline_name != "preprocessing":
-        # with session
+        # with session and with acquisition
         regex_subs.append(
             (
                 (
-                    rf"^{escaped_bids_derivatives_root}/.*?_?session_([^/]+)"
-                    rf"_subject_([^/]+).*/(?:[^/]+)(\.nii(?:\.gz)?)$"
+                    rf"^{escaped_bids_derivatives_root}/.*?_?acquisition_([^/]+)"
+                    rf"_session_([^/]+)_subject_([^/]+).*/"
+                    rf"(?:[^/]+)(\.nii(?:\.gz)?)$"
+                ),
+                (
+                    rf"{bids_derivatives_root}/sub-\3/ses-\2/"
+                    rf"{datatype}/sub-\3_ses-\2_acq-\1_rec-{rec_label}_T2w\4"
+                ),
+            )
+        )
+        # with session and without acquisition
+        regex_subs.append(
+            (
+                (
+                    rf"^{escaped_bids_derivatives_root}/(?!.*?_?acquisition_[^/]+)"
+                    rf".*?_?session_([^/]+)_subject_([^/]+).*/"
+                    rf"(?:[^/]+)(\.nii(?:\.gz)?)$"
                 ),
[…]
