### 2026-08-03 da6055f
add acquisition tag in documentation
