+++ fetpype/utils/utils_bids.py
@@ -257,7 +257,8 @@ def create_bids_datasink(
                         rf"^{escaped_bids_derivatives_root}/"
                         rf"(?!.*?_session_[^/]+).*?_subject_"
                         rf"([^/]+).*/?_cropping.*/"
-                        rf"(sub-[^_]+(?:_acq-[^_]+)?(?:_run-\d+)?)_mask(\.nii(?:\.gz)?)$"
+                        rf"(sub-[^_]+(?:_acq-[^_]+)?(?:_run-\d+)?)_"
+                        rf"mask(\.nii(?:\.gz)?)$"
                     ),
                     (
                         rf"{bids_derivatives_root}/sub-\1/{datatype}"
@@ -272,9 +273,9 @@ def create_bids_datasink(
         regex_subs.append(
             (
                 (
-                    rf"^{escaped_bids_derivatives_root}/.*?_?acquisition_([^/]+)"
-                    rf"_session_([^/]+)_subject_([^/]+).*/"
-                    rf"(?:[^/]+)(\.nii(?:\.gz)?)$"
+                    rf"^{escaped_bids_derivatives_root}/"
+                    rf".*?_?acquisition_([^/]+)_session_([^/]+)"
+                    rf"_subject_([^/]+).*/(?:[^/]+)(\.nii(?:\.gz)?)$"
                 ),
                 (
                     rf"{bids_derivatives_root}/sub-\3/ses-\2/"
@@ -286,7 +287,8 @@ def create_bids_datasink(
         regex_subs.append(
             (
                 (
-                    rf"^{escaped_bids_derivatives_root}/(?!.*?_?acquisition_[^/]+)"
+                    rf"^{escaped_bids_derivatives_root}/"
+                    rf"(?!.*?_?acquisition_[^/]+)"
                     rf".*?_?session_([^/]+)_subject_([^/]+).*/"
                     rf"(?:[^/]+)(\.nii(?:\.gz)?)$"
                 ),
@@ -300,7 +302,8 @@ def create_bids_datasink(
         regex_subs.append(
             (
                 (
-                    rf"^{escaped_bids_derivatives_root}/(?!.*?_?session_[^/]+)"
+                    rf"^{escaped_bids_derivatives_root}/"
+                    rf"(?!.*?_?session_[^/]+)"
                     rf".*?_?acquisition_([^/]+)_subject_([^/]+).*/"
                     rf"(?:[^/]+)(\.nii(?:\.gz)?)$"
                 ),
@@ -314,9 +317,9 @@ def create_bids_datasink(
         regex_subs.append(
             (
                 (
-                    rf"^{escaped_bids_derivatives_root}/(?!.*?_?acquisition_[^/]+)"
-                    rf"(?!.*?_?session_[^/]+).*?_?subject_"
-                    rf"([^/]+).*/(?:[^/]+)(\.nii(?:\.gz)?)$"
+                    rf"^{escaped_bids_derivatives_root}/"
+                    rf"(?!.*?_?acquisition_[^/]+)(?!.*?_?session_[^/]+)"
+                    rf".*?_?subject_([^/]+).*/(?:[^/]+)(\.nii(?:\.gz)?)$"
                 ),
                 (
                     rf"{bids_derivatives_root}/sub-\1/"
@@ -332,12 +335,14 @@ def create_bids_datasink(
             (
                 (
                     rf"^{escaped_bids_derivatives_root}/"
-                    rf".*?_?acquisition_([^/]+)_session_([^/]+)_subject_([^/]+).*/"
+                    rf".*?_?acquisition_([^/]+)_session_([^/]+)"
+                    rf"_subject_([^/]+).*/"
                     rf"input_srr-mask-brain_bounti-19(\.nii(?:\.gz)?)$"
                 ),
[…]
