### 2026-08-06 c0f1328
add doc for singularity
