### 2026-08-28 875eb71
add regex sub for multibounti
